requirements.txt
Flask
Flask-SQLAlchemy
sqlalchemy
cryptography
pymysql
python-dotenv